requirements.txt
paho-mqtt==1.6.1
bottle==0.12.23